View organisation refactor
