+++ src/View/graphicquadrant.cpp
